Get the name of the folder for caching downloaded IMGTHLA files

This function will:

Get the folder name from getOption("hlabud_dir") or else automatically choose an appropriate folder for your operating system thanks to rappdirs.

Create the folder automatically if it does not already exist.

Set the the hlabud_dir option to that new folder.

Usage

get_hlabud_dir()

Value

The name of the folder.

Details

Here are the locations of the hlabud_dir folder on each operating system.

Linux:

~/.local/share/hlabud

Mac:

~/Library/Application Support/hlabud

Windows:

C:\Documents and Settings\{User}\Application Data\slowkow\hlabud

To set the hlabud_dir option, please use:

options(hlabud_dir = "/my/favorite/path")

Examples

if (FALSE) { # \dontrun{
hlabud_dir <- get_hlabud_dir()
} # }