Get HLA frequences from Allele Frequency Net Database (AFND)

Download and read a table of HLA allele frequencies from the Allele Frequency Net Database (AFND).

Usage

hla_frequencies(verbose = FALSE)

Arguments

verbose

If TRUE, print messages along the way.

Value

A dataframe with HLA allele frequencies for all genes.

Details

If you use this data, please cite the latest manuscript about Allele Frequency Net Database:

Gonzalez-Galarza FF, McCabe A, Santos EJMD, Jones J, Takeshita L, Ortega-Rivera ND, et al. Allele frequency net database (AFND) 2020 update: gold-standard data classification, open access genotype data and new query tools. Nucleic Acids Res. 2020;48: D783–D788. doi:10.1093/nar/gkz1029

Examples

# \donttest{
hla_frequencies()
#> # A tibble: 123,502 × 7
#>    group gene  allele  population            indivs_over_n alleles_over_2n     n
#>    <chr> <chr> <chr>   <chr>                         <dbl>           <dbl> <dbl>
#>  1 hla   A     A*01:01 Argentina Rosario To…          15.1           0.076    86
#>  2 hla   A     A*01:01 Armenia combined Reg…          NA             0.125   100
#>  3 hla   A     A*01:01 Australia Cape York …          NA             0.053   103
#>  4 hla   A     A*01:01 Australia Groote Eyl…          NA             0.027    75
#>  5 hla   A     A*01:01 Australia New South …          NA             0.187   134
#>  6 hla   A     A*01:01 Australia Yuendumu A…          NA             0.008   191
#>  7 hla   A     A*01:01 Austria                        27             0.146   200
#>  8 hla   A     A*01:01 Azores Central Islan…          NA             0.08     59
#>  9 hla   A     A*01:01 Azores Oriental Isla…          NA             0.115    43
#> 10 hla   A     A*01:01 Azores Terceira Isla…          NA             0.109   130
#> # ℹ 123,492 more rows
# }