hlabud: Methods for Access and Analysis of the Human Leukocyte Antigen (HLA) Gene Sequence Alignments from IMGTHLA

Fetch sequence alignment data from the IMGTHLA database by Robinson et al (2020) doi:10.1093/nar/gkz950 , automatically convert the sequence alignments into convenient R matrices ready for downstream analysis. The vignette shows examples of using the one-hot encoding of the data with logistic regression and dimensionality reduction. Data is downloaded lazily, as-needed, and cached in a user-configurable folder.

See also

Author

Maintainer: Kamil Slowikowski kslowikowski@gmail.com (ORCID)