Get aligned sequences in a dataframe

Here are the conventions used for alignments (EBI IMGT-HLA):

The entry for each allele is displayed in respect to the reference sequences.

Where identity to the reference sequence is present the base will be displayed as a hyphen (-).

Non-identity to the reference sequence is shown by displaying the appropriate base at that position.

Where an insertion or deletion has occurred this will be represented by a period (.).

If the sequence is unknown at any point in the alignment, this will be represented by an asterisk (*).

In protein alignments for null alleles, the 'Stop' codons will be represented by a hash (X).

In protein alignments, sequence following the termination codon, will not be marked and will appear blank.

These conventions are used for both nucleotide and protein alignments.

Usage

hla_alignments(gene = "DRB1", type = "prot", release = NULL, quiet = TRUE)

Arguments

gene

The name of a gene like "DRB1"

type

The type of sequence, one of "prot", "nuc", "gen"

release

The name of a release like "3.51.0" (optional)

quiet

If FALSE, print messages along the way.

Value

A dataframe.

Examples

# \donttest{
a <- hla_alignments("DRB1")
head(a$sequences)
#>             allele
#> 1 DRB1*01:01:01:01
#> 2 DRB1*01:01:01:02
#> 3 DRB1*01:01:01:03
#> 4 DRB1*01:01:01:04
#> 5 DRB1*01:01:01:05
#> 6 DRB1*01:01:01:06
#>                                                                                                                                                                                                                                                                             seq
#> 1 MVCLKLPGGSCMTALTVTLMVLSSPLALAGDTRPRFLWQLKFECHFFNGTERVR.LLERCIYNQEE.SVRFDSDVGEYRAVTELGRPDAEYWNSQKDLLEQRRAAVDTYCRHNYGVGESFTVQRR.VEPKVTVYPSKTQPLQHHNLLVCSVSGFYPGSIEVRWFRNGQEEKAGVVSTGLIQNGDWTFQTLVMLETVPRSGEVYTCQVEHPSVTSPLTVEWRARSESAQSKMLSGVGGFVLGLLFLGAGLFIYFRNQKGHSGLQPTGFLS
#> 2 ------------------------------------------------------.-----------.----------------------------------------------------------.-----------------------------------------------------------------------------------------------------------------------------------------------
#> 3 ------------------------------------------------------.-----------.----------------------------------------------------------.-----------------------------------------------------------------------------------------------------------------------------------------------
#> 4 ------------------------------------------------------.-----------.----------------------------------------------------------.-----------------------------------------------------------------------------------------------------------------------------------------------
#> 5 ------------------------------------------------------.-----------.----------------------------------------------------------.-----------------------------------------------------------------------------------------------------------------------------------------------
#> 6 ------------------------------------------------------.-----------.----------------------------------------------------------.-----------------------------------------------------------------------------------------------------------------------------------------------
a$aminos[1:6,1:6]
#>                  Pn29 Pn28 Pn27 Pn26 Pn25 Pn24
#> DRB1*01:01:01:01    M    V    C    L    K    L
#> DRB1*01:01:01:02    M    V    C    L    K    L
#> DRB1*01:01:01:03    M    V    C    L    K    L
#> DRB1*01:01:01:04    M    V    C    L    K    L
#> DRB1*01:01:01:05    M    V    C    L    K    L
#> DRB1*01:01:01:06    M    V    C    L    K    L
a$onehot[1:6,1:6]
#>                  Pn29_M Pn28_L Pn28_V Pn27_C Pn26_L Pn25_K
#> DRB1*01:01:01:01      1      0      1      1      1      1
#> DRB1*01:01:01:02      1      0      1      1      1      1
#> DRB1*01:01:01:03      1      0      1      1      1      1
#> DRB1*01:01:01:04      1      0      1      1      1      1
#> DRB1*01:01:01:05      1      0      1      1      1      1
#> DRB1*01:01:01:06      1      0      1      1      1      1
# }