Get a table of allele names for a particular IMGTHLA release

Download a list of all allele names for all HLA genes for a particular IMGTHLA release.

Usage

hla_alleles(release = "latest", overwrite = FALSE, verbose = FALSE)

Arguments

release

Default is "latest". Should be a release name like "3.51.0".

overwrite

Overwrite the existing alleles.json file and Allelelist.{version}.txt file

verbose

If TRUE, print messages along the way.

Value

A data frame with HLA allele ids and names

See also

Examples

# \donttest{
head(hla_alleles())
#>   AlleleID         Allele
#> 1 HLA00001  A*01:01:01:01
#> 2 HLA02169 A*01:01:01:02N
#> 3 HLA14798  A*01:01:01:03
#> 4 HLA15760  A*01:01:01:04
#> 5 HLA16415  A*01:01:01:05
#> 6 HLA16417  A*01:01:01:06
# }